requirements.txt
Pillow
torchmetrics
keras-unet-collection
segmentation-models-pytorch
albumentations
torchinfo
colabcode
pylint
kornia
torchsummary
